refactor(core): move FormatExt files to core/presentation/component/dnd package
