+++ cmp-ttrpg-companion/composeApp/src/commonMain/kotlin/com/cyrillrx/rpg/core/presentation/component/dnd/CreatureFormatExt.kt
@@ -1,4 +1,4 @@
-package com.cyrillrx.rpg.core.presentation.component
+package com.cyrillrx.rpg.core.presentation.component.dnd
 
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.AcUnit
